Extract build checking to module
