+++ lib/deploy/s3/index.js
@@ -1,7 +1,18 @@
-const path = require('path')
-const fs = require('fs-promise')
 const s3 = require('s3')
 const progressBar = require('progress-bar')
+const { checkBuildFiles } = require('../utils')
+
+module.exports = (
+  dir,
+  s3Config
+) => {
+  checkBuildFiles(dir)
+    .then(() => deploy(dir, s3Config))
+    .catch((err) => {
+      console.log(err)
+      process.exit(1)
+    })
+}
 
 const client = s3.createClient({
   maxAsyncS3: 5,
@@ -10,25 +21,6 @@ const client = s3.createClient({
     secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY
   }
 })
-
-module.exports = (
-  dir,
-  s3Config
-) => {
-  Promise.all([
-    checkFileExists(dir, 'index.html'),
-    checkFileExists(dir, 'main.js'),
-    checkFileExists(dir, 'main.css')
-  ]).then(() => deploy(dir, s3Config))
-    .catch((err) => {
-      console.log(err)
-      process.exit(1)
-    })
-}
-
-const checkFileExists = (dir, fileName) => (
-  fs.open(path.join(dir, fileName), 'r')
-)
 
 const deploy = (dir, {
   bucket,
